Move lk_acquire/lk_release to <sys/lock.h>
